requirements.txt
Django>=4.2
djangorestframework
firebase-admin
python-dotenv
psycopg2-binary
django-cors-headers
gunicorn
whitenoise
